[DPEDE-5188] Create pending e2e tests for sidenav component to improve sonarqube in chi repo
